requirements.txt
Django>=3.2.3,<3.3
djangorestframework==3.12.4
easy-thumbnails==2.7.1
social-auth-core==4.1.0
social-auth-app-django==4.0.0
psycopg2>=2.8.6,<2.9
dj-database-url==0.5.0
gunicorn==20.1.0
django-crispy-forms==1.12.0
django-cors-headers==3.7.0
django-cleanup==5.2.0
whitenoise==5.3.0
django-storages==1.11.1
dropbox==11.17.0
coreapi==2.3.3
flake8==3.9.2